+++ packages/semcom-core/lib/component/services/query.service.ts
@@ -1,3 +1,6 @@
+/**
+ * A service to query components.
+ */
 export interface QueryService<T> {
   query(filter: Partial<T>): Promise<T[]>;
 }