+++ packages/semcom-node/lib/store/services/component-pod-store.service.ts
@@ -3,13 +3,27 @@ import { Parser, Quad } from 'n3';
 import { ComponentTransformerService } from '../../component/services/component-transformer.service';
 import { ComponentStore } from './component-store.service';
 
+/**
+ * A { ComponentStore } that stores and retrieves component metadata in/from a pod.
+ */
 export class ComponentPodStore extends ComponentStore {
 
+  /**
+   * Creates a { ComponentPodStore }.
+   *
+   * @param { string } uri - The URI of the pod.
+   * @param { ComponentTransformerService } transformer - Service used to transform components.
+   */
   constructor(
     private readonly uri: string,
     private readonly transformer: ComponentTransformerService
   ) { super(); }
 
+  /**
+   * Fetches all quads from the URI of the pod.
+   *
+   * @returns A list of quads retrieved from the webid.
+   */
   private async fetchAllQuads(): Promise<Quad[]> {
 
     const response = await fetch(this.uri, { headers: { 'Accept': 'text/turtle' } });
@@ -26,6 +40,13 @@ export class ComponentPodStore extends ComponentStore {
 
   }
 
+  /**
+   * Fetch the specified quads from the URI of the pod.
+   *
+   * @param { string } metadataName - The name of the component to fetch.
+   * @returns List of quads retrieved baesed on the metadata name.
+   * @
+   */
   private async fetchQuads(metadataName: string): Promise<Quad[]> {
 
     const response = await fetch(`${this.uri}${metadataName}`, { headers: { 'Accept': 'text/turtle' } });
@@ -35,6 +56,11 @@ export class ComponentPodStore extends ComponentStore {
 
   }
 
+  /**
+   * Transforms all fetched quads to components.
+   *
+   * @returns List of all transformed components.
+   */
   async all(): Promise<ComponentMetadata[]> {
 
     const quads = await this.fetchAllQuads();